fix: Change file names under Story page
