+++ src/pages/story/index.tsx
@@ -1,30 +1,30 @@
-import { NextPage } from "next";
-import React from "react";
-import styled from "styled-components";
-import Footer from "../../common/global/Footer";
-import Section1 from "../../component/story/Section1";
-import Section2 from "../../component/story/Section2";
-import Section3 from "../../component/story/Section3";
-import Section4 from "../../component/story/Section4";
-import Section6 from "../../component/story/Section6";
-import Section7 from "../../component/story/Section7";
+import { NextPage } from 'next';
+import React from 'react';
+import styled from 'styled-components';
+import Footer from '../../common/global/Footer';
+import StoryFarmReview from '../../component/story/StoryFarmReview';
+import StoryHeader from '../../component/story/StoryHeader';
+import StoryProcessCards from '../../component/story/StoryProcessCards';
+import StoryVideo from '../../component/story/StoryVideo';
+import StoryWhoWeAre from '../../component/story/StoryWhoWeAre';
+import StoryHistory from '../../component/story/StoryHistory';
 
 export const StoryContainer = styled.div`
-  width: 100%;
-  background: ${(props) => props.theme.alt.backgroud1};
+	width: 100%;
+	background: ${(props) => props.theme.alt.backgroud1};
 `;
 const Story: NextPage = () => {
-  return (
-    <StoryContainer>
-      <Section1></Section1>
-      <Section2></Section2>
-      <Section3></Section3>
-      <Section4></Section4>
-      <Section6></Section6>
-      <Section7></Section7>
-      <Footer></Footer>
-    </StoryContainer>
-  );
+	return (
+		<StoryContainer>
+			<StoryHeader></StoryHeader>
+			<StoryProcessCards></StoryProcessCards>
+			<StoryVideo></StoryVideo>
+			<StoryFarmReview></StoryFarmReview>
+			<StoryWhoWeAre></StoryWhoWeAre>
+			<StoryHistory></StoryHistory>
+			<Footer></Footer>
+		</StoryContainer>
+	);
 };
 
 export default Story;
